fixed sprite location startup bug
loc and loc2 were not both getting initialized.  This left a dangling pointer
that would sometimes damage screen memory on a restart, it could manifest itself
as a dot missing on the maze.  This was noticed when continually reloading
from vice and probably not so much from a cold machine boot